+++ src/Libraries/SmartStore.Services/Logging/ILogger.cs
@@ -40,7 +40,7 @@ namespace SmartStore.Services.Logging
         /// <param name="pageSize">Page size</param>
         /// <returns>Log item collection</returns>
         IPagedList<Log> GetAllLogs(DateTime? fromUtc, DateTime? toUtc, 
-            string message, LogLevel? logLevel, int pageIndex, int pageSize);
+            string message, LogLevel? logLevel, int pageIndex, int pageSize, int minFrequency);
 
         /// <summary>
         /// Gets a log item