+++ src/Libraries/SmartStore.Services/Logging/NullLogger.cs
@@ -47,7 +47,7 @@ namespace SmartStore.Services.Logging
         /// <param name="pageSize">Page size</param>
         /// <returns>Log item collection</returns>
         public IPagedList<Log> GetAllLogs(DateTime? fromUtc, DateTime? toUtc,
-            string message, LogLevel? logLevel, int pageIndex, int pageSize)
+			string message, LogLevel? logLevel, int pageIndex, int pageSize, int minFrequency)
         {
             return new PagedList<Log>(new List<Log>(), pageIndex, pageSize);
         }